move test in separate package
